service/requirements.txt
requests==2.10.0
xmltodict-0.11.0
